+++ mod-source-record-storage-server/src/main/java/org/folio/services/util/parser/lexeme/operand/ValueBinaryOperand.java
@@ -4,6 +4,7 @@ import org.folio.services.util.parser.lexeme.Lexicon;
 
 import static java.lang.String.format;
 import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_EQUALS;
+import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_IS;
 import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_LEFT_ANCHORED_EQUALS;
 import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_NOT_EQUALS;
 
@@ -12,6 +13,8 @@ import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_NOT_
  * 008.value = '830419m19559999gw mua'    - simple equality
  * 008.value ^= '830419m1'                - left-anchored equality
  * 008.value not= '830419m19559999gw mua' - not equals
+ * 008.value is 'present'                 - check for presence
+ * 008.value is 'absent'                  - check for absence
  */
 public class ValueBinaryOperand extends BinaryOperandLexeme {
 
@@ -26,15 +29,16 @@ public class ValueBinaryOperand extends BinaryOperandLexeme {
   @Override
   public String toSqlRepresentation() {
     StringBuilder stringBuilder = new StringBuilder();
-    String[] keyParts = getKey().split("\\.");
-    String prefix = stringBuilder.append("\"").append("i").append(keyParts[0]).append("\"").append(".\"")
-      .append(keyParts[1]).append("\"").toString();
+    String field = getKey().split("\\.")[0];
+    String prefix = stringBuilder.append("\"").append("i").append(field).append("\".\"value\"").toString();
     if (BINARY_OPERATOR_LEFT_ANCHORED_EQUALS.equals(getOperator())) {
       return prefix + " like ?";
     } else if (BINARY_OPERATOR_EQUALS.equals(getOperator())) {
       return prefix + " = ?";
     } else if (BINARY_OPERATOR_NOT_EQUALS.equals(getOperator())) {
       return stringBuilder.append(" <> ?").toString();
+    } else if (BINARY_OPERATOR_IS.equals(getOperator())) {
+      return PresenceBinaryOperand.getSqlRepresentationForMarcField(field, value);
     }
     throw new IllegalArgumentException(format("Operator [%s] is not supported for the given Value operand", getOperator().getSearchValue()));
   }